Backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.7
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==24.2.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.6
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.3.0
cryptography==44.0.0
cv==1.0.0
cycler==0.12.1
darr==0.6.0
DateTime==5.5
debugpy==1.8.14
decorator==5.2.1
deepface==0.0.93
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.1
dlib==19.24.6
docopt==0.6.2
et_xmlfile==2.0.0
executing==2.2.0
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema==2.21.1
filelock==3.15.4
fire==0.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.27.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
imutils==0.5.4
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.context==6.0.1
jax==0.4.35
jaxlib==0.4.35
jdk==0.0.6
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
Js2Py==0.74
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.1.3
ml_dtypes==0.5.1
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.0
mtcnn==1.0.0
multidict==6.1.0
namex==0.0.8
narwhals==1.40.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
noisereduce==3.0.3
notebook==7.4.1
notebook_shim==0.2.4
numba==0.61.0
numpy==2.0.2
openai==0.28.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipwin==0.5.2
platformdirs==4.2.2
playsound==1.3.0
playwright==1.50.0
plotly==5.24.1
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pvporcupine==3.0.5
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydicom==3.0.1
pyee==12.1.1
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.10.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.0
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pyttsx3==2.98
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retina-face==0.0.17
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.1
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.28.1
Send2Trash==1.8.3
setuptools==76.0.0
SimpleITK==2.4.1
simsimd==6.2.1
six==1.16.0
sklearn-compat==0.1.3
smmap==5.0.2
smote==0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sound==0.1.0
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
SpeechRecognition==3.11.0
split-folders==0.5.1
SQLAlchemy==2.0.37
sqlparse==0.5.1
stack-data==0.6.3
streamlit==1.45.1
stringzilla==3.12.5
sympy==1.13.1
task-automation==0.0.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.19.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
timm==1.0.15
tinycss2==1.4.0
tk==0.1.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
transformers==4.47.1
trio==0.28.0
trio-websocket==0.11.1
twilio==9.5.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3
ultralytics==8.3.90
ultralytics-thop==2.0.14
unity==8.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.3
wakeonlan==3.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
whisper==1.1.10
wikipedia==1.4.0
Wikipedia-API==0.8.1
wolframalpha==5.1.3
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.14.2
yarl==1.18.3
zope.interface==7.2
